# Resolved Conflicts:
#	src/fusion.cpp
#	src/include/miopen/fusion.hpp
#	src/kernels/gfx900_56.cd.pdb.txt
#	src/kernels/gfx900_64.cd.pdb.txt
#	src/kernels/gfx906_56.cd.pdb.txt
#	src/kernels/gfx906_64.cd.pdb.txt


[ROCm/MIOpen commit: 25c11c634e2f5cf3cf5f58ca4f28defa815c2d53]

+++ src/include/miopen/util.hpp
@@ -38,7 +38,7 @@ namespace miopen {
 float Im2ColGPU(Handle& handle,
                 int data_size,
                 ConstData_t im,
-                size_t im_offset,
+                int im_offset,
                 int c,
                 int h,
                 int w,
@@ -70,7 +70,7 @@ float Col2ImGPU(Handle& handle,
                 int h,
                 int w,
                 Data_t im,
-                size_t im_offset);
+                int im_offset);
 
 float transpose_NCHW2CNHW(Handle& handle,
                           int n,
